+++ route/route.go
@@ -1,4 +1,4 @@
-package router
+package route
 
 import (
 	"errors"
